+++ src/main/java/org/folio/search/service/ResourceService.java
@@ -26,7 +26,6 @@ import java.util.stream.Stream;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
 import org.apache.commons.collections4.CollectionUtils;
-import org.folio.search.configuration.properties.SearchConfigurationProperties;
 import org.folio.search.domain.dto.FolioIndexOperationResponse;
 import org.folio.search.domain.dto.ResourceEvent;
 import org.folio.search.domain.dto.ResourceEventType;
@@ -40,6 +39,7 @@ import org.folio.search.repository.IndexRepository;
 import org.folio.search.repository.PrimaryResourceRepository;
 import org.folio.search.repository.ResourceRepository;
 import org.folio.search.service.consortium.ConsortiumInstanceService;
+import org.folio.search.service.consortium.ConsortiumTenantService;
 import org.folio.search.service.converter.MultiTenantSearchDocumentConverter;
 import org.folio.search.service.metadata.ResourceDescriptionService;
 import org.folio.search.utils.SearchUtils;
@@ -60,7 +60,7 @@ public class ResourceService {
   private final ResourceDescriptionService resourceDescriptionService;
   private final MultiTenantSearchDocumentConverter multiTenantSearchDocumentConverter;
   private final Map<String, ResourceRepository> resourceRepositoryBeans;
-  private final SearchConfigurationProperties searchConfig;
+  private final ConsortiumTenantService consortiumTenantService;
   private final TenantScopedExecutionService tenantScopedExecutionService;
   private final ConsortiumInstanceService consortiumInstanceService;
 
@@ -77,9 +77,7 @@ public class ResourceService {
       return getSuccessIndexOperationResponse();
     }
 
-    var eventsToIndex = searchConfig.inConsortiaMode()
-      ? resources
-      : getEventsThatCanBeIndexed(resources, SearchUtils::getIndexName);
+    var eventsToIndex = getEventsToIndex(resources);
     var elasticsearchDocuments = multiTenantSearchDocumentConverter.convert(eventsToIndex);
     return indexSearchDocuments(elasticsearchDocuments);
   }
@@ -97,9 +95,7 @@ public class ResourceService {
       return getSuccessIndexOperationResponse();
     }
 
-    var eventsToIndex = searchConfig.inConsortiaMode()
-      ? resourceIdEvents
-      : getEventsThatCanBeIndexed(resourceIdEvents, SearchUtils::getIndexName);
+    var eventsToIndex = getEventsToIndex(resourceIdEvents);
 
     var eventsByTenant = eventsToIndex.stream()
       .filter(event -> INSTANCE_RESOURCE.equals(event.getResourceName()))
@@ -124,6 +120,15 @@ public class ResourceService {
       getNumberOfRequests(indexDocuments), getNumberOfRequests(removeDocuments), getErrorMessage(bulkIndexResponse));
 
     return bulkIndexResponse;
+  }
+
+  private List<ResourceEvent> getEventsToIndex(List<ResourceEvent> events) {
+    var inConsortium = events.stream()
+      .map(ResourceEvent::getTenant)
+      .distinct()
+      .anyMatch(tenantId -> consortiumTenantService.getCentralTenant(tenantId).isPresent());
+
+    return inConsortium ? events : getEventsThatCanBeIndexed(events, SearchUtils::getIndexName);
   }
 
   private Map<String, List<SearchDocumentBody>> processIndexInstanceEvents(String tenant,